backend 2/requirements.txt
# backend/requirements.txt

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0    # PostgreSQL
pymysql>=1.0.0            # MySQL 
pyodbc>=4.0.0             # MSSQL 

# Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Async
asyncio
aiofiles

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# simulation
simpy==4.1.1
numpy==1.26.2

# etc.
asyncpg==0.29.0
aioredis==2.0.1
websockets==12.0
python-multipart==0.0.6
redis==5.0.1
rich
watchdog